conformance/basicoperations/cl_program_build.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLProgram build conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script id="square" type="x-kernel">
__kernel void square(
    __global float* input,
    __global float* output,
    const unsigned int count)
{
    int i = get_global_id(0);
    if(i < count)
        output[i] = input[i] * input[i];
}
</script>
<script>
function callback(event, userData) {
    testPassed("event status : " + event.status + "UserData = " + userData);
    notifyFinishedToHarness();
}

try {
    debug("WebCLProgram :: build() test case.\n");
    debug("(WebCLDevice[] devices, optional DOMString? options, optional WebCLCallback whenFinished, optional any userdata)\n");
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var kernelSource = wtu.getKernel("square");
    var webCLProgram = wtu.createProgram(webCLContext, kernelSource);
    var webCLPlatforms = wtu.getPlatforms();


    var webCLDeviceGPU = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_GPU);
    shouldBeTrue("wtu.build(webCLProgram, webCLDeviceGPU);");
    var webCLDeviceALL = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_ALL);
    shouldBeTrue("wtu.build(webCLProgram, webCLDeviceALL);");
    var webCLDeviceACCELERATOR = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_ACCELERATOR);
    shouldBeTrue("wtu.build(webCLProgram, webCLDeviceACCELERATOR);");
    var webCLDeviceDEFAULT = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_DEFAULT);
    shouldBeTrue("wtu.build(webCLProgram, webCLDeviceDEFAULT);");
    var webCLDeviceCPU = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_CPU);
    shouldBeTrue("wtu.build(webCLProgram, webCLDeviceCPU);");
    debug("");

    // Testing for build-options.
    var webCLDevice = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_CPU);

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-single-precision-constant\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-denorms-are-zero\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-opt-disable\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-mad-enable\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-no-signed-zeroes\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-unsafe-math-optimizations\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-finite-math-only\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-fast-relaxed-math\");");

    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-strict-aliasing\");");

    // Testing with callback and userData.
    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-single-precision-constant\", callback);");
    shouldBeTrue("wtu.build(webCLProgram, webCLDevice, \"-cl-single-precision-constant\", callback, \"test\");");

    // Testing for Exception.
    shouldThrowExceptionName("wtu.build(webCLProgram, invalid_object);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.build(webCLProgram, webCLDevice, invalid_object);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.build(webCLProgram, webCLDevice, \"-cl-no-signed-zeroes\", invalid_object);", "INVALID_VALUE");

} catch (e) {
    testFailed("WebCLProgram :: build () throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
